+++ .travis.yml
@@ -48,12 +48,12 @@ before_script:
   - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd ../..; fi
 script:
   # Run tests (linux only due to docker support... )
-  - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd $TRAVIS_BUILD_DIR && python3 tests/integration_tests/flows/_test_mariadb.py; fi
-  - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd $TRAVIS_BUILD_DIR && python3 tests/integration_tests/flows/_test_clickhouse.py; fi
+  - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd $TRAVIS_BUILD_DIR && python3 tests/integration_tests/flows/test_mariadb.py; fi
+  - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd $TRAVIS_BUILD_DIR && python3 tests/integration_tests/flows/test_clickhouse.py; fi
   - if [[ "$TRAVIS_OS_NAME" = "linux" && $(lsof -t -i:47334) ]]; then kill -9 $(lsof -t -i:47334); fi
+  - if [ "$TRAVIS_OS_NAME" = "linux" ]; then ps -aux | grep $(lsof -t -i:47334) || true; fi
   - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd $TRAVIS_BUILD_DIR/tests/integration_tests/api && python3 test_http.py; fi
-  # - if [ "$TRAVIS_OS_NAME" = "linux" ]; then cd $TRAVIS_BUILD_DIR/tests/integration_tests/flows; fi
-  # - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sleep 61 && python3 test_clickhouse.py; fi
+
   # Build windows installer
   - mkdir -p build;
   - if [ "$TRAVIS_OS_NAME" = "windows" ]; then cd $TRAVIS_BUILD_DIR/distributions/windows; fi
