+++ mindsdb/api/http/start.py
@@ -1,9 +1,7 @@
-import json
 import os
 import mindsdb
 import logging
 import sys
-import random
 
 from mindsdb.api.http.namespaces.predictor import ns_conf as predictor_ns
 from mindsdb.api.http.namespaces.datasource import ns_conf as datasource_ns
@@ -11,15 +9,12 @@ from mindsdb.api.http.namespaces.util import ns_conf as utils_ns
 from mindsdb.api.http.initialize import initialize_flask, initialize_interfaces
 from mindsdb.utilities.config import Config
 
+
 def start(config, initial=False):
     if not initial:
         print('\n\nWarning, this process should not have been started... nothing is "wrong" but it needlessly ate away a tiny bit of precious comute !\n\n')
     config = Config(config)
-    port=47334
-    host='0.0.0.0'
-    debug=False
-
-
+    debug = False
 
     if not logging.root.handlers:
         rootLogger = logging.getLogger()
@@ -32,14 +27,13 @@ def start(config, initial=False):
         errStream.addFilter(lambda record: record.levelno > logging.INFO)
         rootLogger.addHandler(errStream)
 
-    mindsdb.CONFIG.MINDSDB_DATASOURCES_PATH = os.path.join(mindsdb.CONFIG.MINDSDB_STORAGE_PATH,'datasources')
-    mindsdb.CONFIG.MINDSDB_TEMP_PATH = os.path.join(mindsdb.CONFIG.MINDSDB_STORAGE_PATH,'tmp')
+    mindsdb.CONFIG.MINDSDB_DATASOURCES_PATH = os.path.join(mindsdb.CONFIG.MINDSDB_STORAGE_PATH, 'datasources')
+    mindsdb.CONFIG.MINDSDB_TEMP_PATH = os.path.join(mindsdb.CONFIG.MINDSDB_STORAGE_PATH, 'tmp')
 
     os.makedirs(mindsdb.CONFIG.MINDSDB_STORAGE_PATH, exist_ok=True)
     os.makedirs(mindsdb.CONFIG.MINDSDB_DATASOURCES_PATH, exist_ok=True)
     os.makedirs(mindsdb.CONFIG.MINDSDB_TEMP_PATH, exist_ok=True)
-    #'''
-    
+
     app, api = initialize_flask(config)
     initialize_interfaces(config, app)
 
@@ -47,7 +41,9 @@ def start(config, initial=False):
     api.add_namespace(datasource_ns)
     api.add_namespace(utils_ns)
 
+    print(f"Start on {config['api']['http']['host']}:{config['api']['http']['port']}")
     app.run(debug=debug, port=config['api']['http']['port'], host=config['api']['http']['host'])
+
 
 if __name__ == '__main__':
     start()
