+++ tests/integration_tests/flows/test_clickhouse.py
@@ -1,41 +1,53 @@
-from subprocess import Popen
-import time
-import os
-import signal
-import random
-import string
-import requests
-
 import unittest
+import requests
+import os
+import csv
+import inspect
+import subprocess
+import atexit
 
+from mindsdb.interfaces.native.mindsdb import MindsdbNative
 from mindsdb.utilities.config import Config
-import mindsdb
 
-rand = ''.join(random.choice(string.ascii_uppercase) for _ in range(11))
-ds_name = f'default.hr_ds_{rand}'
-pred_name = f'hr_predictor_{rand}'
+from common import (
+    wait_api_ready,
+    prepare_config,
+    wait_db,
+    is_container_run,
+    TEST_CONFIG,
+    TESTS_ROOT,
+    OUTPUT
+)
 
-# Can't be a fixture since it's used in setup/teardown
-root = 'http://localhost:47334'
+TEST_CSV = {
+    'name': 'home_rentals.csv',
+    'url': 'https://s3.eu-west-2.amazonaws.com/mindsdb-example-data/home_rentals.csv'
+}
+TEST_DATA_TABLE = 'home_rentals'
+TEST_PREDICTOR_NAME = 'test_predictor'
 
-def set_get_config_path():
-    os.environ['DEV_CONFIG_PATH'] = 'config'
-    return os.environ['DEV_CONFIG_PATH'] + '/config.json'
+config = Config(TEST_CONFIG)
 
-def query_ch(query, database='default'):
-    config = Config(set_get_config_path())
-    add = ' FORMAT JSON'
-    for ele in ['drop ', 'create ','insert ', 'show ']:
-        if ele in query.lower():
-            add = ''
-    query += add
 
-    connect_string = 'http://{}:{}'.format(
-        config['integrations']['default_clickhouse']['host'],
-        config['integrations']['default_clickhouse']['port']
-    )
+def query(query):
+    if 'CREATE ' not in query.upper() and 'INSERT ' not in query.upper():
+        query += ' FORMAT JSON'
 
-    params = {'user': config['integrations']['default_clickhouse']['user'], 'password': config['integrations']['default_clickhouse']['password'], 'database': database}
+    host = config['integrations']['default_clickhouse']['host']
+    port = config['integrations']['default_clickhouse']['port']
+
+    connect_string = f'http://{host}:{port}'
+
+    params = {'user': 'default'}
+    try:
+        params['user'] = config['integrations']['default_clickhouse']['user']
+    except Exception:
+        pass
+
+    try:
+        params['password'] = config['integrations']['default_clickhouse']['password']
+    except Exception:
+        pass
 
     res = requests.post(
         connect_string,
@@ -44,101 +56,226 @@ def query_ch(query, database='default'):
     )
 
     if res.status_code != 200:
-        print(f'Error in query: {query}')
+        print(f'ERROR: code={res.status_code} msg={res.text}')
+        raise Exception()
 
     if ' FORMAT JSON' in query:
-        return res.json()['data']
-    else:
-        return res.text
+        res = res.json()['data']
+
+    return res
+
 
 class ClickhouseTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        confpath = set_get_config_path()
-        c = Config(confpath)
-        if os.path.isdir(c['interface']['datastore']['storage_dir']) is False:
-            os.makedirs(c['interface']['datastore']['storage_dir'], exist_ok=True)
-        if os.path.isdir(c['interface']['mindsdb_native']['storage_dir']) is False:
-            os.makedirs(c['interface']['mindsdb_native']['storage_dir'], exist_ok=True)
+        cls.mdb = MindsdbNative(config)
 
-        cls.sp = Popen(['python3', '-m', 'mindsdb'], close_fds=True)
+        models = cls.mdb.get_models()
+        models = [x['name'] for x in models]
+        if TEST_PREDICTOR_NAME in models:
+            cls.mdb.delete_model(TEST_PREDICTOR_NAME)
 
-        for i in range(20):
-            try:
-                res = requests.get(f'{root}/util/ping')
-                if res.status_code != 200:
-                    raise Exception('')
-            except:
-                time.sleep(1)
-                if i == 19:
-                    raise Exception("Can't connect !")
+        query('create database if not exists test')
+        test_tables = query('show tables from test')
+        test_tables = [x['name'] for x in test_tables]
+        if TEST_DATA_TABLE not in test_tables:
+            print('creating test data table...')
+            query(f'''
+                CREATE TABLE test.{TEST_DATA_TABLE} (
+                    id Int16,
+                    number_of_rooms Int8,
+                    number_of_bathrooms Int8,
+                    sqft Int32,
+                    location String,
+                    days_on_market Int16,
+                    initial_price Int32,
+                    neighborhood String,
+                    rental_price Int32
+                ) ENGINE = MergeTree()
+                ORDER BY id
+                PARTITION BY location
+            ''')
 
-        #query_ch('DROP DATABASE mindsdb')
-        query_ch(f"""
-        CREATE TABLE {ds_name} (number_of_rooms String, number_of_bathrooms String, sqft Int64, location String, days_on_market Int64, initial_price Int64, neighborhood String, rental_price Float64)  ENGINE=URL('https://raw.githubusercontent.com/mindsdb/mindsdb-examples/master/benchmarks/home_rentals/dataset/train.csv', CSVWithNames)
+            test_csv_path = str(TESTS_ROOT.joinpath('temp/', TEST_CSV['name']).resolve())
+            if os.path.isfile(test_csv_path) is False:
+                r = requests.get(TEST_CSV['url'])
+                with open(test_csv_path, 'wb') as f:
+                    f.write(r.content)
+
+            with open(test_csv_path) as f:
+                csvf = csv.reader(f)
+                i = 0
+                for row in csvf:
+                    if i > 0:
+                        number_of_rooms = int(row[0])
+                        number_of_bathrooms = int(row[1])
+                        sqft = int(float(row[2].replace(',', '.')))
+                        location = str(row[3])
+                        days_on_market = int(row[4])
+                        initial_price = int(row[5])
+                        neighborhood = str(row[6])
+                        rental_price = int(float(row[7]))
+                        query(f'''INSERT INTO test.{TEST_DATA_TABLE} VALUES (
+                            {i},
+                            {number_of_rooms},
+                            {number_of_bathrooms},
+                            {sqft},
+                            '{location}',
+                            {days_on_market},
+                            {initial_price},
+                            '{neighborhood}',
+                            {rental_price}
+                        )''')
+                    i += 1
+            print('done')
+
+    def test_1_initial_state(self):
+        print(f'\nExecuting {inspect.stack()[0].function}')
+        print('Check all testing objects not exists')
+
+        print(f'Predictor {TEST_PREDICTOR_NAME} not exists')
+        models = [x['name'] for x in self.mdb.get_models()]
+        self.assertTrue(TEST_PREDICTOR_NAME not in models)
+
+        print('Test datasource exists')
+        test_tables = query('show tables from test')
+        test_tables = [x['name'] for x in test_tables]
+        self.assertTrue(TEST_DATA_TABLE in test_tables)
+
+        print('Test predictor table not exists')
+        mindsdb_tables = query('show tables from mindsdb')
+        mindsdb_tables = [x['name'] for x in mindsdb_tables]
+        self.assertTrue(TEST_PREDICTOR_NAME not in mindsdb_tables)
+
+        print('mindsdb.predictors table exists')
+        self.assertTrue('predictors' in mindsdb_tables)
+
+        print('mindsdb.commands table exists')
+        self.assertTrue('commands' in mindsdb_tables)
+
+    def test_2_insert_predictor(self):
+        print(f'\nExecuting {inspect.stack()[0].function}')
+        query(f"""
+            insert into mindsdb.predictors (name, predict_cols, select_data_query, training_options) values
+            (
+                '{TEST_PREDICTOR_NAME}',
+                'rental_price, location',
+                'select * from test.{TEST_DATA_TABLE} limit 100',
+                '{{"join_learn_process": true, "stop_training_in_x_seconds": 3}}'
+            );
         """)
 
-    @classmethod
-    def tearDownClass(cls):
-        try:
-            cls.sp.kill()
-        except:
-            pass
-        # try:
-        #     pgrp = os.getpgid(cls.sp.pid)
-        #     os.killpg(pgrp, signal.SIGINT)
-        #     os.remove(set_get_config_path())
-        #     os.system('fuser -k 47335/tcp ; fuser -k 47334/tcp')
-        # except:
-        #     pass
+        print('predictor record in mindsdb.predictors')
+        res = query(f"select status from mindsdb.predictors where name = '{TEST_PREDICTOR_NAME}'")
+        self.assertTrue(len(res) == 1)
+        self.assertTrue(res[0]['status'] == 'complete')
 
-    def test_1_setup(self):
-        result = query_ch(f"show tables", 'mindsdb')
-        assert 'predictors' in result
-        assert 'commands' in result
+        print('predictor table in mindsdb db')
+        mindsdb_tables = query('show tables from mindsdb')
+        mindsdb_tables = [x['name'] for x in mindsdb_tables]
+        self.assertTrue(TEST_PREDICTOR_NAME in mindsdb_tables)
 
-    def test_2_learn(self):
-        q = f"""
-            insert into mindsdb.predictors
-                (name, predict_cols, select_data_query, training_options)
-            values (
-                '{pred_name}',
-                'rental_price',
-                'SELECT * FROM {ds_name} LIMIT 400',
-                '{{"stop_training_in_x_seconds": 10}}'
-            )
-        """
-        result = query_ch(q)
-        time.sleep(40)
-        for i in range(40):
-            try:
-                result = query_ch(f"SELECT name FROM mindsdb.predictors where name='{pred_name}'")
-                if result[0]['name'] != pred_name:
-                    raise Exception('not ready yet !')
-            except:
-                time.sleep(1)
-                if i == 39:
-                    raise Exception("Can't get predictor !")
+    def test_3_query_predictor(self):
+        print(f'\nExecuting {inspect.stack()[0].function}')
+        res = query(f"""
+            select
+                rental_price, location, sqft, number_of_rooms,
+                rental_price_confidence, rental_price_min, rental_price_max, rental_price_explain
+            from
+                mindsdb.{TEST_PREDICTOR_NAME} where sqft=1000
+        """)
 
-        result = query_ch(f"show tables", 'mindsdb')
-        assert pred_name in result
+        print('check result')
+        self.assertTrue(len(res) == 1)
 
-    def test_3_predict_from_where(self):
-        result = query_ch(f"SELECT rental_price FROM mindsdb.{pred_name} where sqft=1000 and location='good'")
-        assert len(result) == 1
-        assert 'rental_price' in result[0]
+        res = res[0]
 
-    def test_4_predict_from_query(self):
-        len_ds = query_ch(f'SELECT COUNT(*) as len from {ds_name}')[0]['len']
-        result = query_ch(f""" SELECT rental_price, rental_price_explain, rental_price_confidence, rental_price_max, rental_price_min FROM mindsdb.{pred_name} where select_data_query=='SELECT * FROM {ds_name}' """)
+        self.assertTrue(res['rental_price'] is not None and res['rental_price'] != 'None')
+        self.assertTrue(res['location'] is not None and res['location'] != 'None')
+        # NOTE in current Clickhouse all int fields returns as strings
+        self.assertTrue(res['sqft'] == '1000')
+        self.assertIsInstance(res['rental_price_confidence'], float)
+        self.assertIsInstance(res['rental_price_min'], int)
+        self.assertIsInstance(res['rental_price_max'], int)
+        self.assertIsInstance(res['rental_price_explain'], str)
+        self.assertTrue(res['number_of_rooms'] == 'None' or res['number_of_rooms'] is None)
 
-        assert int(len(result)) == int(len_ds)
-        for res in result:
-            assert 'rental_price' in res
-            assert 'rental_price_explain' in res
-            assert 'rental_price_confidence' in res
-            assert 'rental_price_max' in res
-            assert 'rental_price_min' in res
+    def test_4_range_query(self):
+        print(f'\nExecuting {inspect.stack()[0].function}')
 
-if __name__ == '__main__':
-    unittest.main()
+        results = query(f"""
+            select
+                rental_price, location, sqft, number_of_rooms,
+                rental_price_confidence, rental_price_min, rental_price_max, rental_price_explain
+            from
+                mindsdb.{TEST_PREDICTOR_NAME} where select_data_query='select * from test.{TEST_DATA_TABLE} limit 3'
+        """)
+
+        print('check result')
+        self.assertTrue(len(results) == 3)
+        for res in results:
+            self.assertTrue(res['rental_price'] is not None and res['rental_price'] != 'None')
+            self.assertTrue(res['location'] is not None and res['location'] != 'None')
+            self.assertIsInstance(res['rental_price_confidence'], float)
+            self.assertIsInstance(res['rental_price_min'], int)
+            self.assertIsInstance(res['rental_price_max'], int)
+            self.assertIsInstance(res['rental_price_explain'], str)
+
+    def test_5_delete_predictor_by_command(self):
+        print(f'\nExecuting {inspect.stack()[0].function}')
+
+        query(f"""
+            insert into mindsdb.commands values ('delete predictor {TEST_PREDICTOR_NAME}');
+        """)
+
+        print(f'Predictor {TEST_PREDICTOR_NAME} not exists')
+        models = [x['name'] for x in self.mdb.get_models()]
+        self.assertTrue(TEST_PREDICTOR_NAME not in models)
+
+        print('Test predictor table not exists')
+        mindsdb_tables = query('show tables from mindsdb')
+        mindsdb_tables = [x['name'] for x in mindsdb_tables]
+        self.assertTrue(TEST_PREDICTOR_NAME not in mindsdb_tables)
+
+
+def stop_clickhouse():
+    ch_sp = subprocess.Popen(
+        ['./cli.sh', 'clickhouse-stop'],
+        cwd=TESTS_ROOT.joinpath('docker/').resolve(),
+        stdout=OUTPUT,
+        stderr=OUTPUT
+    )
+    ch_sp.wait()
+
+
+if __name__ == "__main__":
+    temp_config_path = prepare_config(config, 'default_clickhouse')
+
+    if is_container_run('clickhouse-test') is False:
+        subprocess.Popen(
+            ['./cli.sh', 'clickhouse'],
+            cwd=TESTS_ROOT.joinpath('docker/').resolve(),
+            stdout=OUTPUT,
+            stderr=OUTPUT
+        )
+        atexit.register(stop_clickhouse)
+    clickhouse_ready = wait_db(config, 'default_clickhouse')
+
+    if clickhouse_ready:
+        sp = subprocess.Popen(
+            ['python3', '-m', 'mindsdb', '--api', 'mysql', '--config', temp_config_path],
+            stdout=OUTPUT,
+            stderr=OUTPUT
+        )
+        atexit.register(sp.kill)
+
+    api_ready = clickhouse_ready and wait_api_ready(config)
+
+    try:
+        if api_ready is False:
+            print(f'Failed by timeout. ClickHouse started={clickhouse_ready} MindsDB started={api_ready}')
+            raise Exception()
+        unittest.main(failfast=True)
+        print('Tests passed!')
+    except Exception as e:
+        print(f'Tests Failed!\n{e}')
